Transferred vocabulary import logic to new dedicated dialogue component
