+++ src/styles/EditorDialog.sass
@@ -114,7 +114,7 @@ div.add-new-vocab {
   width: 100%;
 }
 
-div.vocabulary-editor {
+div.vocabulary-editor, div.external-vocabulary-import {
 
   .form-group {
     .control-label {
@@ -145,6 +145,15 @@ div.vocabulary-editor {
     }
   }
 }
+
+div.external-vocabulary-import {
+  padding: 1em;
+  div.external-vocabulary-import-properties {
+    width: 60%;
+    margin-left: 2em;
+  }
+}
+
 
 #ccrTable {
   margin-bottom: 30px;
